% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcount.R
\name{amen_count_nb}
\alias{amen_count_nb}
\title{AMEN Count Model with Negative Binomial Distribution}
\usage{
amen_count_nb(
  data,
  X,
  ndim = 2,
  niter = 15000,
  nburn = 2500,
  nthin = 5,
  nprint = 500,
  pr_sd_z = 0.5,
  pr_sd_w = 0.5,
  jump_beta = 0.4,
  jump_alpha = 1,
  jump_gamma = 0.025,
  jump_z = 0.5,
  jump_w = 0.5,
  jump_eta = 1,
  jump_delta = 1,
  eta_init = 0,
  r_init = 1,
  pr_mean_beta = 0,
  pr_sd_beta = 1,
  pr_mean_alpha = 0,
  pr_sd_alpha = 1,
  jump_r = 0.025,
  pr_mean_r = 0.5,
  pr_sd_r = 1,
  pr_mean_delta = 0,
  pr_sd_delta = 1,
  pr_mean_gamma = 0.5,
  pr_sd_gamma = 1,
  pr_a_beta = 0.001,
  pr_b_beta = 0.001,
  pr_a_alpha = 0.001,
  pr_b_alpha = 0.001,
  pr_a_eta = 1,
  pr_b_eta = 1,
  direct = TRUE,
  overdispersion = FALSE,
  zeroinflate = FALSE,
  fix = FALSE,
  fixsd = FALSE,
  missing = -99,
  verbose = FALSE,
  single = FALSE,
  singledist = TRUE,
  vector = FALSE,
  test = FALSE,
  fix_r = FALSE,
  covariate = FALSE,
  hierarchical_r = FALSE,
  pr_a_r_sd = 0.001,
  pr_b_r_sd = 0.001,
  skip = FALSE
)
}
\arguments{
\item{data}{A matrix or data frame containing count response data}

\item{X}{Design matrix for covariates}

\item{ndim}{Number of latent dimensions (default: 2)}

\item{niter}{Number of MCMC iterations (default: 15000)}

\item{nburn}{Number of burn-in iterations (default: 2500)}

\item{nthin}{Thinning interval (default: 5)}

\item{nprint}{Print interval for progress (default: 500)}

\item{pr_sd_z}{Prior standard deviation for z parameters (default: 0.5)}

\item{pr_sd_w}{Prior standard deviation for w parameters (default: 0.5)}

\item{jump_beta}{Jump size for beta parameters (default: 0.4)}

\item{jump_alpha}{Jump size for alpha parameters (default: 1)}

\item{jump_gamma}{Jump size for gamma parameter (multiplicative weight term) (default: 0.025)}

\item{jump_z}{Jump size for z parameters (default: 0.5)}

\item{jump_w}{Jump size for w parameters (default: 0.5)}

\item{jump_eta}{Jump size for eta parameters (default: 1)}

\item{jump_delta}{Jump size for delta parameters (default: 1)}

\item{eta_init}{Initial value for eta (default: 0.0)}

\item{r_init}{Initial value for r parameter (default: 1)}

\item{pr_mean_beta}{Prior mean for beta parameters (default: 0)}

\item{pr_sd_beta}{Prior standard deviation for beta parameters (default: 1)}

\item{pr_mean_alpha}{Prior mean for alpha parameters (default: 0)}

\item{pr_sd_alpha}{Prior standard deviation for alpha parameters (default: 1)}

\item{jump_r}{Jump size for r parameter (default: 0.025)}

\item{pr_mean_r}{Prior mean for r parameter (default: 0.5)}

\item{pr_sd_r}{Prior standard deviation for r parameter (default: 1)}

\item{pr_mean_delta}{Prior mean for delta parameters (default: 0)}

\item{pr_sd_delta}{Prior standard deviation for delta parameters (default: 1)}

\item{pr_mean_gamma}{Prior mean for gamma parameter (multiplicative weight term) (default: 0.5)}

\item{pr_sd_gamma}{Prior standard deviation for gamma parameter (multiplicative weight term) (default: 1)}

\item{pr_a_beta}{Prior shape parameter for beta (default: 0.001)}

\item{pr_b_beta}{Prior rate parameter for beta (default: 0.001)}

\item{pr_a_alpha}{Prior shape parameter for alpha (default: 0.001)}

\item{pr_b_alpha}{Prior rate parameter for alpha (default: 0.001)}

\item{pr_a_eta}{Prior shape parameter for eta (default: 1)}

\item{pr_b_eta}{Prior rate parameter for eta (default: 1)}

\item{direct}{Logical for direct sampling (default: TRUE)}

\item{overdispersion}{Logical for overdispersion (default: FALSE)}

\item{zeroinflate}{Logical for zero inflation (default: FALSE)}

\item{fix}{Logical for fixed parameters (default: FALSE)}

\item{fixsd}{Logical for fixed standard deviation (default: FALSE)}

\item{missing}{Missing value indicator (default: -99)}

\item{verbose}{Logical for verbose output (default: FALSE)}

\item{single}{Logical for single chain (default: FALSE)}

\item{singledist}{Logical for single distribution (default: TRUE)}

\item{vector}{Logical for vector operations (default: FALSE)}

\item{test}{Logical for test mode (default: FALSE)}

\item{fix_r}{Logical for fixed r parameter (default: FALSE)}

\item{covariate}{Logical for covariate inclusion (default: FALSE)}

\item{hierarchical_r}{Logical for hierarchical r (default: FALSE)}

\item{pr_a_r_sd}{Prior shape for r standard deviation (default: 0.001)}

\item{pr_b_r_sd}{Prior rate for r standard deviation (default: 0.001)}

\item{skip}{Logical to skip certain computations (default: FALSE)}
}
\value{
A list containing:
\item{data}{Original data matrix}
\item{bic}{Bayesian Information Criterion}
\item{waic}{Widely Applicable Information Criterion}
\item{dic}{Deviance Information Criterion}
\item{beta_estimate}{Posterior mean estimates for additive row effects}
\item{alpha_estimate}{Posterior mean estimates for additive column effects}
\item{gamma_estimate}{Posterior mean estimate for multiplicative weight parameter}
\item{r_estimate}{Posterior mean estimate for overdispersion parameter}
\item{z_estimate}{Posterior mean estimates for row latent positions}
\item{w_estimate}{Posterior mean estimates for column latent positions}
\item{mcmc_inf}{MCMC information including burn-in, iterations, and thinning}
\item{map_inf}{Maximum a posteriori information}
\item{beta}{Full posterior samples for additive row effects}
\item{alpha}{Full posterior samples for additive column effects}
\item{gamma}{Full posterior samples for multiplicative weight parameter}
\item{r}{Full posterior samples for overdispersion parameter}
\item{z}{Procrustes-aligned posterior samples for row latent positions}
\item{w}{Procrustes-aligned posterior samples for column latent positions}
\item{accept_*}{Acceptance rates for various parameters}
}
\description{
Fits an Additive and Multiplicative Effects Network (AMEN) model for count data 
using negative binomial distribution with MCMC algorithm
}
\examples{
\dontrun{
# Example usage with real data
# AL_exp_d should be your count data matrix
# X should be your design matrix

model_exp_neg <- amen_count_nb(data=AL_exp_d, X = X, niter=100000, nburn=10000,
                               nthin = 5, overdispersion = TRUE, zeroinflate = FALSE,
                               jump_r = 0.5, r_init = 1,
                               verbose = FALSE, direct = TRUE, fix = TRUE, fixsd = TRUE,
                               jump_gamma = 0.5, jump_beta = 2.5, jump_alpha = 2.5,
                               jump_z = 0.6, jump_w = 0.6, pr_sd_z = 1, pr_sd_w = 1,
                               pr_sd_beta = 3, pr_sd_alpha = 3,
                               pr_mean_r = log(1), pr_sd_r = 0.15,
                               eta_init = 0.5, pr_a_eta = 1, pr_b_eta = 1, jump_eta = 0.3,
                               pr_mean_gamma = 0.5, pr_sd_gamma = 1, ndim = 2,
                               singledist = FALSE, single = FALSE, hierarchical_r = FALSE,
                               vector = TRUE, missing = -99, covariate = FALSE, fix_r = FALSE)

}

}
